Setup bors-ng
Add support for using bors to do pre-merge testing. In particular avoid
running CI on branches ending in .tmp as bors uses those for
preparation and add a new dummy job to key the bors status against to
avoid having to call out all test jobs specifically which would be hard
to maintain.

Closes: #18

Signed-off-by: Rafael Garcia Ruiz <rafael.garcia@collabora.com>