Support copying a remote image without a bmap file
Handles the remote nobmap option. copy_remote_input_nobmap function
replicates a lot of code from the original copy_remote_input.
On the other hand avoids the use of bmap and has a spinner instead of
a progress bar since we can't know the total size of the disk image.

Signed-off-by: Rafael Garcia Ruiz <rafael.garcia@collabora.com>