bmap-parser: require Send bounds on traits
Not having the Send bound on the AsyncSeekForward trait in principle
makes it easier to match. However it limits where the async copy can be
used quite a bit. In particular it means it can't be used in a
tokio::spawn, which is rather a big disadvantage.

So start Send for the trait...

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>